+++ src/Http/Controllers/InitiatePaymentController.php
@@ -3,6 +3,7 @@
 namespace Naykel\Payit\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Naykel\Payit\Models\PaymentPlatform;
 use Naykel\Payit\PaymentPlatformResolver;
 
 class InitiatePaymentController
@@ -11,23 +12,20 @@ class InitiatePaymentController
 
     public function __invoke(Request $request)
     {
-        // Validate the incoming request for platform selection and agreement.
-        $request->validate([
-            'platformId' => ['required'],
-            'agree' => 'accepted',
-            // payment_method is a value returned from the stripe.js client-side
-            // script. It is required only when the platform ID is 2 (Stripe).
-            // This will be reviewed in a future iteration as it is not a good
-            // practice to have this validation and hardcoded value here.
-            'payment_method' => 'required_if:platformId,2',
+        $platformIds = PaymentPlatform::query()->pluck('id')->all();
+
+        $validated = $request->validate([
+            'platformId' => ['required', 'integer', 'in:' . implode(',', $platformIds)],
+            'agree' => ['accepted'],
+            'payment_method' => ['string', 'required_if:platformId,2,12'],
         ]);
 
         // Resolve the payment service using the provided platform ID.
         // e.g. StripeService or PayPalService
-        $service = $this->paymentPlatformResolver->resolveService($request->platformId);
+        $service = $this->paymentPlatformResolver->resolveService($validated['platformId']);
 
         // store the selected platform ID in the session for later use.
-        session()->put('payment.ppid', $request->platformId);
+        session()->put('payment.ppid', $validated['platformId']);
 
         // Use the resolved service to handle the payment process.
         return $service->initiatePayment(session('cart.total'), $request);