+++ src/Http/Controllers/PaymentSuccessController.php
@@ -8,7 +8,7 @@ class PaymentSuccessController
     {
         // NK::TD Refactor to allow for more flexibility and reusability. Maybe
         // return a payment success dto.
-        return redirect()->route('user.dashboard')
+        return redirect()->route(config('payit.return_route'))
             ->withSuccess('Your payment has been processed');
     }
 }