+++ src/routes.php
@@ -7,7 +7,6 @@ use Naykel\Payit\Http\Controllers\PaymentCancelledController;
 use Naykel\Payit\Http\Controllers\PaymentSuccessController;
 
 Route::middleware(['web', 'auth'])->prefix('payment')->name('payment.')->group(function () {
-
     Route::post('/initiate', InitiatePaymentController::class)->name('initiate');
     Route::get('/approval', PaymentApprovalController::class)->name('approval');
     Route::get('/success', PaymentSuccessController::class)->name('success');